Refactor Anikoto and AnimeSogo to use new multisrc anikototheme
