+++ lib-multisrc/anikototheme/build.gradle.kts
@@ -1,0 +1,7 @@
+import keiyoushi.gradle.extensions.baseVersionCode
+
+plugins {
+    alias(kei.plugins.multisrc)
+}
+
+baseVersionCode = 1
