+++ lib-multisrc/anikototheme/src/main/java/eu/kanade/tachiyomi/multisrc/anikototheme/AnikotoThemeDto.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.animeextension.en.animesogo
+package eu.kanade.tachiyomi.multisrc.anikototheme
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
