+++ lib-multisrc/anikototheme/src/main/java/eu/kanade/tachiyomi/multisrc/anikototheme/EpisodeMetadataFetcher.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.animeextension.all.anikoto
+package eu.kanade.tachiyomi.multisrc.anikototheme
 
 import android.util.Log
 import kotlinx.serialization.SerialName
